+++ init.sh
@@ -1,0 +1,4 @@
+cd ~
+export PATH="$PATH:`pwd`/flutter/bin"
+which flutter
+cd /workspaces/APP_FreshPlus_TakeCareMyRefrigerator
