+++ run.sh
@@ -1,0 +1,1 @@
+flutter run -d web-server --web-hostname=0.0.0.0
